游戏操作自动化 (1.1.3)   12/20 ~ 1/30 目标：实战—自动迷雾寻路
@1. UIManager.ShowPanel() 可以实时指定界面位置。
@2. 做输出日志弹窗(大类上限100条)，实装，能catch错误信息并输出堆栈。debug能力↑
@3. 做变量监控弹窗，实装。debug能力↑
@4.
"键盘节点"：改成Update中进行，按下、松开
"等待节点"：新增，它只负责等待
"地图寻路节点"：新增，它只负责寻路并赋值方向变量。
@5. 寻路，封装路径为BigCellPath。寻路输出方向—由两个大格子的终点遍历到起点，选连线间无障碍的。
@6. 优化DrawProcessPanel体验。用户可以手动设置节点的槽点位置，从而让连线看的整齐。
@7. 实战：编写“测试移动“脚本，探索脚本编写规律，提升与3个debug窗的配合默契。目前脚本能自动探索迷雾

+++ Assets/Script/UI/Panel/Auto/AutoPanel.cs
@@ -215,7 +215,7 @@ namespace Script.UI.Panel.Auto
 
                 bool repeat = repeatKeySet.Contains(key);
                 // AU.PostMessagePacked(selectedWin, key, repeat);
-                WU.keybd_event_packed((int)key);
+                WU.keybd_event_packed(key);
                 repeatKeySet.Add(key);
             }
             else
@@ -228,7 +228,7 @@ namespace Script.UI.Panel.Auto
                 // else
                 // {
                 // AU.PostMessagePacked(selectedWin, key, true, false);
-                WU.keybd_event_packed((int)key, false);
+                WU.keybd_event_packed(key, false);
                 repeatKeySet.Remove(key);
                 // }
